starknet_provider: move `ResourceBounds` into `components`
