+++ packages/starknet_provider/lib/src/model/components/resource_bounds.freezed.dart
